requirements.txt
openai==0.27.5
streamlit
python-dotenv
db-sqlite3
